requirements.txt
fastapi
uvicorn
chromadb
sentence-transformers
transformers
rank-bm25
langchain
langchain-community
langchain-core
google-generativeai
pandas
openpyxl